requirements.txt
pandas 
numpy
matplotlib.pyplot
pynance
seaborn
scipy
pyti
talib
math
string
nltk
sklearn
